requirements.txt
requests
PyGithub
python-telegram-bot